// src/app/core/helios-api-client/model/statutMembreDTO.ts
/**
 * Helios | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NullableOfStatutsMembres } from './nullableOfStatutsMembres';


export interface StatutMembreDTO { 
    id: number | null;
    libelle: string | null;
    code: NullableOfStatutsMembres | null;
}
export namespace StatutMembreDTO {
}